pyproject.toml
[project]
name = "fuelmate-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ctorum",email = "dev.rezendevitor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
